+++ raw-data-listener/src/main/java/org/example/StreamableEntitiesProducer.java
@@ -4,18 +4,16 @@ import java.time.Instant;
 import java.util.Optional;
 import java.util.UUID;
 
-import lombok.extern.slf4j.Slf4j;
 import org.example.events.ExtractedAisMessage;
 import org.springframework.context.event.EventListener;
 import org.springframework.kafka.core.KafkaTemplate;
 import org.springframework.stereotype.Service;
 
 import com.ais.avro.schemas.AisMessage;
-import com.ais.avro.schemas.ConnectionStatus;
-import com.ais.avro.schemas.Heartbeat;
 import com.ais.avro.schemas.InterfaceEvent;
 
 import lombok.AllArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 
 @Service
 @AllArgsConstructor
@@ -27,7 +25,8 @@ public class StreamableEntitiesProducer {
   @EventListener
   public void sendMessage(ExtractedAisMessage extractedAisMessage) {
     getAisAvroMessage(extractedAisMessage.getAisMessage())
-            .ifPresent(aisMessage -> {
+        .ifPresent(
+            aisMessage -> {
               produceMessage(InterfaceEvent.newBuilder().setAisMessage(aisMessage));
               log.info("Produces converted ais message, MMSI: {}", aisMessage.getMmsi());
             });
@@ -46,15 +45,17 @@ public class StreamableEntitiesProducer {
       return Optional.empty();
     }
 
-    if (aisMessage.getValidPosition() == null || aisMessage.getValidPosition().getLatitude() == 0 || aisMessage.getValidPosition().getLongitude() == 0) {
+    if (aisMessage.getValidPosition() == null
+        || aisMessage.getValidPosition().getLatitude() == 0
+        || aisMessage.getValidPosition().getLongitude() == 0) {
       log.warn("Got a message with invalid position. MMSI: {}, rejecting", aisMessage.getUserId());
     }
 
-    return Optional.of(AisMessage.newBuilder()
-                    .setMmsi(aisMessage.getUserId())
-                    .setLatitude(aisMessage.getValidPosition().getLatitude())
-                    .setLongitude(aisMessage.getValidPosition().getLongitude())
-           .build());
+    return Optional.of(
+        AisMessage.newBuilder()
+            .setMmsi(aisMessage.getUserId())
+            .setLatitude(aisMessage.getValidPosition().getLatitude())
+            .setLongitude(aisMessage.getValidPosition().getLongitude())
+            .build());
   }
-
 }
