+++ raw-data-listener/src/main/java/org/example/events/ExtractedAisMessage.java
@@ -1,14 +1,16 @@
 package org.example.events;
 
+import org.springframework.context.ApplicationEvent;
+
 import dk.dma.ais.message.AisMessage;
 import lombok.Getter;
-import org.springframework.context.ApplicationEvent;
 
 @Getter
 public class ExtractedAisMessage extends ApplicationEvent {
-    private final AisMessage aisMessage;
-    public ExtractedAisMessage(Object source, AisMessage aisMessage) {
-        super(source);
-        this.aisMessage = aisMessage;
-    }
+  private final AisMessage aisMessage;
+
+  public ExtractedAisMessage(Object source, AisMessage aisMessage) {
+    super(source);
+    this.aisMessage = aisMessage;
+  }
 }
